Update pretty much all dependencies, add Cargo.lock files to git (#3856)
* Update a lot of dependencies

* Add cago lockfiles to repo

* fixed critical whitespace

* fixes

* lockfiles

* Fix wide blend deprecation

Replace with newer select

* fmt

* fix fixes

* increase msrv

* fixes

* fix CI

* update z3, fix nostd

* more z3?

* more z3?

* actions

* Error handling and trying CI again

* CI GODS PLZ

* link arg

* more fix?

* z3

* fuuuu z3

* z3

---------

+++ crates/libafl_qemu/src/modules/utils/addr2line.rs
@@ -1,4 +1,4 @@
-#![allow(clippy::unnecessary_cast)]
+#![allow(clippy::unnecessary_cast, clippy::cast_sign_loss)]
 //! Utils for addr2line
 
 use std::{borrow::Cow, fmt::Write, fs};
@@ -30,7 +30,7 @@ pub fn is_pie(file: object::File<'_>) -> bool {
             if let Ok(Some(d)) = dyn_sec {
                 let arr = d.0;
                 for v in arr {
-                    if v.d_tag.get(elf.endian()) == DT_FLAGS_1
+                    if v.d_tag.get(elf.endian()) as u64 == DT_FLAGS_1
                         && v.d_val.get(elf.endian()) & DF_1_PIE == DF_1_PIE
                     {
                         is_pie = true;
